fix: fixing a bunch of issues with cache and shortcuts
